fix texture
