+++ src/main/java/de/melanx/utilitix/content/bell/ItemMobBell.java
@@ -1,17 +1,13 @@
 package de.melanx.utilitix.content.bell;
 
-import de.melanx.utilitix.UtilitiX;
 import de.melanx.utilitix.util.MobUtil;
 import io.github.noeppi_noeppi.libx.mod.ModX;
-import net.minecraft.ChatFormatting;
 import net.minecraft.nbt.Tag;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.MutableComponent;
-import net.minecraft.network.chat.TranslatableComponent;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.entity.MobCategory;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.SpawnEggItem;
@@ -23,7 +19,6 @@ import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import java.util.List;
 import java.util.Objects;
-import java.util.Optional;
 
 public class ItemMobBell extends BellBase {
 
@@ -51,25 +46,8 @@ public class ItemMobBell extends BellBase {
     @Override
     public void appendHoverText(@Nonnull ItemStack stack, @Nullable Level level, @Nonnull List<Component> tooltip, @Nonnull TooltipFlag flag) {
         super.appendHoverText(stack, level, tooltip, flag);
-        MutableComponent component = getCurrentMob(stack);
+        MutableComponent component = MobUtil.getCurrentMob(stack);
         tooltip.add(component != null ? component : MobUtil.NO_MOB);
-    }
-
-    @Nullable
-    public static MutableComponent getCurrentMob(ItemStack stack) {
-        String s = stack.getOrCreateTag().getString(MobUtil.ENTITY_TYPE_TAG);
-        Optional<EntityType<?>> entityType = EntityType.byString(s);
-
-        return entityType.map(ItemMobBell::getCurrentMob).orElse(null);
-    }
-
-    @Nonnull
-    public static MutableComponent getCurrentMob(EntityType<?> entityType) {
-        Component name = entityType.getDescription();
-        MutableComponent component = new TranslatableComponent("tooltip." + UtilitiX.getInstance().modid + ".current_mob");
-        component.withStyle(entityType.getCategory() == MobCategory.MONSTER ? ChatFormatting.RED : ChatFormatting.GOLD);
-
-        return component.append(": ").append(name);
     }
     
     public static int getColor(ItemStack stack) {
