+++ src/main/java/de/melanx/utilitix/item/ItemMobYoinker.java
@@ -1,9 +1,9 @@
 package de.melanx.utilitix.item;
 
 import de.melanx.utilitix.UtilitiX;
-import de.melanx.utilitix.content.bell.ItemMobBell;
 import de.melanx.utilitix.util.MobUtil;
 import io.github.noeppi_noeppi.libx.base.ItemBase;
+import net.minecraft.client.renderer.item.ItemProperties;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.MutableComponent;
@@ -16,16 +16,25 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.item.context.UseOnContext;
 import net.minecraft.world.level.Level;
+import net.minecraftforge.client.IItemRenderProperties;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import java.util.List;
 import java.util.Optional;
+import java.util.function.Consumer;
 
 public class ItemMobYoinker extends ItemBase {
 
+    private static final String TAG_FILLED = "filled";
+
     public ItemMobYoinker(Properties properties) {
         super(UtilitiX.getInstance(), properties);
+    }
+
+    @Override
+    public void initializeClient(@Nonnull Consumer<IItemRenderProperties> consumer) {
+        ItemProperties.register(this, UtilitiX.getInstance().resource("filled"), ((stack, level, entity, seed) -> stack.getOrCreateTag().getBoolean(TAG_FILLED) ? 1.0F : 0.0F));
     }
 
     @Nonnull
@@ -64,9 +73,15 @@ public class ItemMobYoinker extends ItemBase {
     }
 
     @Override
+    public void inventoryTick(@Nonnull ItemStack stack, @Nonnull Level level, @Nonnull Entity entity, int slotId, boolean isSelected) {
+        boolean filled = MobUtil.getCurrentMob(stack) != null;
+        stack.getOrCreateTag().putBoolean(TAG_FILLED, filled);
+    }
+
+    @Override
     public void appendHoverText(@Nonnull ItemStack stack, @Nullable Level level, @Nonnull List<Component> tooltip, @Nonnull TooltipFlag flag) {
         super.appendHoverText(stack, level, tooltip, flag);
-        MutableComponent component = ItemMobBell.getCurrentMob(stack);
+        MutableComponent component = MobUtil.getCurrentMob(stack);
         tooltip.add(component != null ? component : MobUtil.NO_MOB);
     }
