+++ src/main/resources/assets/utilitix/textures/item/mob_yoinker_filled.png.mcmeta
@@ -1,0 +1,6 @@
+{
+  "animation": {
+    "interpolate": true,
+    "frametime": 3
+  }
+}
